Rename `oneLineCollapseWhitespace` to `renderOneLine`
