+++ packages/11ty/_plugins/transforms/README.md
@@ -1,6 +1,6 @@
 This directory includes templating helpers to transform text beyond functions included by `common-tags` 
 
-## `oneLineCollapseWhitespace`
+## `renderOneLine`
 
 Similar to `oneLine`, this function flattens html tags into a single line, stripping out whitespace that surrounds text nodes inside of tags. It allows template HTML for elements to include indentation for better readability without rendering the extra whitespace.
 
@@ -20,7 +20,7 @@ Gets rendered like this:
 <span> Some inline text </span>
 ```
 
-Instead, `oneLineCollapseWhitespace` renders it like this:
+Instead, `renderOneLine` renders it like this:
 
 ```html
 <span>Some inline text</span>
